Moved last method of Utils to CrySLRule class
After removing the last method in Utils and adding it to CrySLRule class, the Utils class was deleted since it was redundant.
